# monoes_apis/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Web Scraper API",
    "description": "API for generating XPath configurations and extracting data from HTML using Gemini AI.",
    "version": "1.0.0"
  },
  "paths": {
    "/generate-config": {
      "post": {
        "summary": "Generate Config",
        "operationId": "generate_config_generate_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/extract": {
      "post": {
        "summary": "Extract",
        "operationId": "extract_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/extracttest": {
      "post": {
        "summary": "Extract Test",
        "operationId": "extract_test_extracttest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/configs": {
      "get": {
        "summary": "List Configs",
        "operationId": "list_configs_configs_get",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by config name",
              "title": "Filter"
            },
            "description": "Filter by config name"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/configs/{config_name}": {
      "get": {
        "summary": "Get Config",
        "operationId": "get_config_configs__config_name__get",
        "parameters": [
          {
            "name": "config_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExtractRequest": {
        "properties": {
          "htmlContent": {
            "type": "string",
            "title": "Htmlcontent"
          },
          "configName": {
            "type": "string",
            "title": "Configname"
          }
        },
        "type": "object",
        "required": [
          "htmlContent",
          "configName"
        ],
        "title": "ExtractRequest"
      },
      "ExtractTestRequest": {
        "properties": {
          "htmlContent": {
            "type": "string",
            "title": "Htmlcontent"
          },
          "configName": {
            "type": "string",
            "title": "Configname"
          }
        },
        "type": "object",
        "required": [
          "htmlContent",
          "configName"
        ],
        "title": "ExtractTestRequest"
      },
      "ExtractionField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExtractionField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "type"
        ],
        "title": "ExtractionField"
      },
      "ExtractionSchema": {
        "properties": {
          "fields": {
            "$ref": "#/components/schemas/ExtractionField"
          }
        },
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "ExtractionSchema"
      },
      "GenerateConfigRequest": {
        "properties": {
          "htmlContent": {
            "type": "string",
            "title": "Htmlcontent"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "extractionSchema": {
            "$ref": "#/components/schemas/ExtractionSchema"
          },
          "configName": {
            "type": "string",
            "title": "Configname"
          }
        },
        "type": "object",
        "required": [
          "htmlContent",
          "purpose",
          "extractionSchema",
          "configName"
        ],
        "title": "GenerateConfigRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}